feat(notifier): refactor edit provider
