+++ src/components/pages/notifier/NotifierRoutes.tsx
@@ -8,15 +8,15 @@ import Logger from 'utils/Logger'
 import TabsTemplate from 'components/templates/TabsTemplate'
 import { getTabValueFromLocation } from 'utils/utils'
 import { NotifierOffersContextProvider } from 'context/Services/notifier/offers'
+import ManageProviderPage from 'components/pages/notifier/provider/ManageProviderPage'
+import WithLoginCard from 'components/hoc/WithLoginCard'
 import NotifierLandingPage from './NotifierLandingPage'
 import { NotFound } from '..'
-import NotifierSellPage from './sell/NotifierSellPage'
 import NotifierOfferCheckoutPage from './buy/NotifierOfferCheckoutPage'
 import NotifierOffersPage from './buy/NotifierOffersPage'
 import NotifierMyOffersPage from './myoffers/NotifierMyOffersPage'
 import NotifierMyPurchasePage from './mypurchase/NotifierMyPurchasePage'
 import { buildTabs } from '../routerUtils'
-import NotifierEditOffersPage from './myoffers/NotifierEditOffersPage'
 
 const logger = Logger.getInstance()
 
@@ -92,7 +92,7 @@ const NotifierRoutes: FC = () => {
                   <Route
                     exact
                     path={ROUTES.NOTIFIER.MYOFFERS.EDIT}
-                    component={NotifierEditOffersPage}
+                    component={ManageProviderPage}
                   />
 
                   {DeadEndRoute}
@@ -104,7 +104,14 @@ const NotifierRoutes: FC = () => {
             <Route
               exact
               path={ROUTES.NOTIFIER.SELL.BASE}
-              component={NotifierSellPage}
+              component={
+        WithLoginCard({
+          WrappedComponent: () => <ManageProviderPage create />,
+          title: 'Connect your wallet to register as a provider',
+          contentText: 'Please, connect your wallet in order to register as a provider.',
+        })
+
+              }
             />
 
             {/* My purchases */}
