--- src/components/pages/notifier/myoffers/NotifierEditOffersPage.tsx
@@ -1,6 +1,0 @@
-import React, { FC } from 'react'
-import ProviderEdit from 'components/organisms/notifier/offers/ProviderEdit'
-
-const NotifierEditOffersPage: FC = () => <ProviderEdit mode="EDIT" />
-
-export default NotifierEditOffersPage
